+++ src/Driver/tester.py
@@ -1,4 +1,0 @@
-import sys
-
-plat = sys.platform.lower()
-print(plat)